requirements/requirements_linux.txt
opencv_contrib_python_headless==4.2.0.34
numpy==1.19.0
scipy==1.5.0
pandas==1.1.2
Pillow==8.0.1

